third_party/blink/web_tests/fast/dom/shadow/touch-event-expected.txt

CONSOLE WARNING: line 15: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Touch event should be fired in Shadow DOM.

Elements in ShadowDOM should not be revealed in touchTarget if it's examined in non shadow tree.

PASS touchstartFiredInShadowDOM is true
PASS touchTargetForHost1 is host1
PASS touchTargetForInput1 is input1
PASS successfullyParsed is true

TEST COMPLETE
